libs/api/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paint-crud-api"
version = "0.1.0"
description = "CRUD API service for paint catalog management"
authors = [{name = "mrcndz"}]
dependencies = [
    "fastapi>=0.104.0",
    "fastapi-cors>=0.0.6",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "psycopg2-binary>=2.9.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    "pydantic>=2.4.0",
    "pydantic-settings>=2.0.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "shared",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
